+++ app/src/main/java/ch/epfl/cs311/wanderwave/model/localDb/AppDatabase.kt
@@ -1,4 +1,4 @@
-package ch.epfl.cs311.wanderwave.model.local
+package ch.epfl.cs311.wanderwave.model.localDb
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
